% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{Pull Surveycto data in R}
\usage{
pull_data(formid, servername, username, password, key = NULL, newserver = T)
}
\arguments{
\item{formid}{character: is the Surveycto form id.}

\item{servername}{character: is the name of the Surveycto server.}

\item{username}{character: is the username for the account being used to login to the server.}

\item{password}{character: is the password for the account being used to login to the server.}

\item{key}{character: name of the private encryption file, if applicable. If it doesn't contain an absolute path, the file name is relative to the current working directory, getwd(). If the form is encrypted and a key is not provided, only the fields marked as publishable will be returned. Defaults to NULL.}

\item{newserver}{logical: TRUE for surveycto version 2.70 and above. FALSE otherwise. Default is TRUE.}
}
\value{
A dataframe containing the formdata in wide format
}
\description{
This function pulls data from surveycto using their api. Typically, numeric fields will be read as character by R when downloaded using the api. The user will manually need to convert the object type as necessary.
}
