% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_media.R
\name{pull_media}
\alias{pull_media}
\title{Save Surveycto media files using R}
\usage{
pull_media(
  df,
  variable,
  username,
  password,
  folder,
  uidvar,
  fileformat,
  key = NULL
)
}
\arguments{
\item{df}{is the name of the dataframe which contains surveycto data downloaded using pull_data.}

\item{variable}{character: is the name of the form variable which contains the media.}

\item{username}{character: is the username for the account being used to login to the server.}

\item{password}{character: is the password for the account being used to login to the server.}

\item{folder}{character: is the absolute path to the folder where the media files will be downloaded.}

\item{uidvar}{character: any unique id from the df dataset. This variable is used to name the media files. This variable shouldn't contain special characters such as (: , ; ') as such characters aren't allowed in the filename.}

\item{fileformat}{character: is the format in which to save the file. Examples - jpg, png, mp3}

\item{key}{character: name of the private encryption file, if applicable. If it doesn't contain an absolute path, the file name is relative to the current working directory, getwd(). If the form is encrypted and a key is not provided, only the fields marked as publishable will be returned. Defaults to NULL.}
}
\value{
The media files corresponding to the required variable will be downloaded in the specified folder.
}
\description{
This function saves the media that were collected during the survey. pull_media should be used once data is downloaded using pull_data and an R dataframe is generated. In the data frame, the media variable will contain links that will be used to download the media files
}
\examples{
pull_media("dataframe_formdata","respondent_picture","user(at)gmail.com","passwordhere","C:/Users//Dropbox/projectfolder/mediafiles/pictures","jpg")
pull_media("dataframe_formdata","respondent_picture","user(at)gmail.com","passwordhere","C:/Users//Dropbox/projectfolder/mediafiles/pictures","jpg",key="C:/Users/Dropbox/projectfolder/encryption_keys/projectdata_PRIVATEDONOTSHARE.pem")
}
