% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_mount.R
\name{vc_mount}
\alias{vc_mount}
\title{Mounts veracrypt container directly from R}
\usage{
vc_mount(
  filename,
  drive,
  password,
  vcexecutable = "C:\\\\Program Files\\\\VeraCrypt\\\\VeraCrypt.exe"
)
}
\arguments{
\item{filename}{character: name of the veracrypt container. If it doesn't contain an absolute path, the file name is relative to the current working directory, getwd().}

\item{drive}{character: drive name. Example - "Z://"}

\item{password}{character: password of the veracrypt container.}

\item{vcexecutable}{character: name of the veracrypt executable file. If it doesn't contain an absolute path, the file name is relative to the current working directory, getwd(). Default is "C:/Program Files/VeraCrypt/VeraCrypt.exe"}
}
\value{
Running the function will open veracrypt as specified. Returns a NULL object in R.
}
\description{
This function mounts the veracrypt container directly from R. Also check the sister function \code{\link{vc_dismount}}.
}
\examples{
vc_mount("C:/User/Dropbox/Project1/raw_data/veracrypt_raw", "k://", "veracryptpassword")

vc_mount(filename="C:/User/Dropbox/Project1/raw_data/veracrypt_raw", drive="k://", password="veracryptpassword", vcexecutable="C:/Program Files (x86)/VeraCrypt/VeraCrypt.exe")

}
