% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_open.R
\name{vc_open}
\alias{vc_open}
\title{Opens veracrypt container directly from R}
\usage{
vc_open(
  filename,
  drive,
  password,
  vcexecutable = "C:\\\\Program Files\\\\VeraCrypt\\\\VeraCrypt.exe"
)
}
\arguments{
\item{filename}{character: name of the veracrypt container. If it doesn't contain an absolute path, the file name is relative to the current working directory, getwd().}

\item{drive}{character: drive name. Example - "Z://"}

\item{password}{character: password of the veracrypt container.}

\item{vcexecutable}{character: name of the veracrypt executable file. Default is "C:/Program Files/VeraCrypt/VeraCrypt.exe"}
}
\value{
Running the function will open veracrypt as specified. Returns a NULL object in R.
}
\description{
This function opens the veracrypt container directly from R
}
