% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_dismount.R
\name{vc_dismount}
\alias{vc_dismount}
\title{Dismounts veracrypt container directly from R}
\usage{
vc_dismount(
  drive,
  vcexecutable = "C:\\\\Program Files\\\\VeraCrypt\\\\VeraCrypt.exe"
)
}
\arguments{
\item{drive}{character: name of the drive from which to dismount the volume. Example - "Z://"}

\item{vcexecutable}{character: name of the veracrypt executable file. If it doesn't contain an absolute path, the file name is relative to the current working directory, getwd(). Default is "C:/Program Files/VeraCrypt/VeraCrypt.exe"}
}
\value{
Running the function will dismount the veracrypt volume as specified. Returns a NULL object in R.
}
\description{
This function dismounts the veracrypt container directly from R. Also check the sister function \code{\link{vc_mount}}.
}
\examples{
vc_dismount("k://")

vc_dismount("k://", vcexecutable="C:/Program Files/Veracrypt/Veracrypt.exe")

vc_dismount("k://", vcexecutable="C:/Program Files (x86)/Veracrypt/Veracrypt.exe")
}
